chatgpt4telegram/requirements.txt
python-telegram-bot==20.3
g4f
